Large refactoring of the system
This change is to prepare the repository for major changes ahead as I build my personal website
This regarding has come from the idea as I move towards creating what is known as "The Lost and Found Archives" which will be the place to find all of the projects and work that I have done, as a way for me to personally track my ventures and to show of the cool things I am working on in an informal way. 

This is not going to hit the ground all polished and perfect but rather to facilitate building things over time. small steady steps